+++ src/main/java/svenhjol/charmonium/base/CharmoniumSounds.java
@@ -2,10 +2,8 @@ package svenhjol.charmonium.base;
 
 import net.minecraft.sound.SoundEvent;
 import net.minecraft.util.Identifier;
-import net.minecraft.util.registry.Registry;
 import svenhjol.charm.base.handler.RegistryHandler;
 import svenhjol.charmonium.Charmonium;
-
 
 import java.util.HashMap;
 import java.util.Map;
@@ -17,9 +15,13 @@ public class CharmoniumSounds {
     public static final SoundEvent MUSIC_THARNA = createSound("music.tharna");
     public static final SoundEvent MUSIC_MUS = createSound("music.mus");
     public static final SoundEvent MUSIC_UNDIR = createSound("music.undir");
+    public static final SoundEvent MUSIC_RUIN = createSound("music.ruin");
     public static final SoundEvent MUSIC_DISC = createSound("music_disc.charmonium");
 
     // ambience
+    public static final SoundEvent AMBIENCE_BADLANDS_DAY_LONG = createSound("ambience.badlands.day_long");
+    public static final SoundEvent AMBIENCE_BADLANDS_DAY_SHORT = createSound("ambience.badlands.day_short");
+    public static final SoundEvent AMBIENCE_BADLANDS_NIGHT_LONG = createSound("ambience.badlands.night_long");
     public static final SoundEvent AMBIENCE_BEACH_LONG = createSound("ambience.beach.long");
     public static final SoundEvent AMBIENCE_CAVE_LONG = createSound("ambience.cave.long");
     public static final SoundEvent AMBIENCE_CAVE_SHORT = createSound("ambience.cave.short");
@@ -31,14 +33,15 @@ public class CharmoniumSounds {
     public static final SoundEvent AMBIENCE_DESERT_NIGHT_SHORT = createSound("ambience.desert.night_short");
     public static final SoundEvent AMBIENCE_END_SHORT = createSound("ambience.end.short");
     public static final SoundEvent AMBIENCE_END_LONG = createSound("ambience.end.long");
-    public static final SoundEvent AMBIENCE_EXTREME_HILLS_DAY_LONG = createSound("ambience.extreme_hills.day_long");
-    public static final SoundEvent AMBIENCE_EXTREME_HILLS_DAY_SHORT = createSound("ambience.extreme_hills.day_short");
-    public static final SoundEvent AMBIENCE_EXTREME_HILLS_NIGHT_LONG = createSound("ambience.extreme_hills.night_long");
+    public static final SoundEvent AMBIENCE_MOUNTAINS_DAY_LONG = createSound("ambience.mountains.day_long");
+    public static final SoundEvent AMBIENCE_MOUNTAINS_DAY_SHORT = createSound("ambience.mountains.day_short");
+    public static final SoundEvent AMBIENCE_MOUNTAINS_NIGHT_LONG = createSound("ambience.mountains.night_long");
     public static final SoundEvent AMBIENCE_FOREST_DAY_LONG = createSound("ambience.forest.day_long");
     public static final SoundEvent AMBIENCE_FOREST_NIGHT_LONG = createSound("ambience.forest.night_long");
     public static final SoundEvent AMBIENCE_ICY_DAY_LONG = createSound("ambience.icy.day_long");
     public static final SoundEvent AMBIENCE_ICY_DAY_SHORT = createSound("ambience.icy.day_short");
     public static final SoundEvent AMBIENCE_ICY_NIGHT_LONG = createSound("ambience.icy.night_long");
+    public static final SoundEvent AMBIENCE_ICY_THUNDERSTORM = createSound("ambience.icy.thunderstorm");
     public static final SoundEvent AMBIENCE_JUNGLE_DAY_LONG = createSound("ambience.jungle.day_long");
     public static final SoundEvent AMBIENCE_JUNGLE_NIGHT_LONG = createSound("ambience.jungle.night_long");
     public static final SoundEvent AMBIENCE_MINESHAFT_SHORT = createSound("ambience.mineshaft.short");