+++ src/main/java/svenhjol/charmonium/client/MusicClient.java
@@ -2,19 +2,27 @@ package svenhjol.charmonium.client;
 
 import net.minecraft.util.Identifier;
 import net.minecraft.world.biome.Biome;
+import svenhjol.charm.base.CharmClientModule;
+import svenhjol.charm.base.CharmModule;
 import svenhjol.charm.base.handler.ModuleHandler;
-import svenhjol.charm.client.MusicClient;
-import svenhjol.charm.client.MusicCondition;
-import svenhjol.charmonium.base.CharmoniumSounds;
 import svenhjol.charm.base.helper.DimensionHelper;
+import svenhjol.charm.client.MusicImprovementsClient;
+import svenhjol.charm.client.MusicImprovementsClient.MusicCondition;
+import svenhjol.charm.client.PlayerStateClient;
+import svenhjol.charmonium.base.CharmoniumSounds;
 
-public class AmbientMusicClient {
-    public AmbientMusicClient() {
+public class MusicClient extends CharmClientModule {
+    public MusicClient(CharmModule module) {
+        super(module);
+    }
+
+    @Override
+    public void register() {
         if (!ModuleHandler.enabled("charm:music_improvements"))
             return;
 
         // play Þarna in overworld anywhere
-        MusicClient.getMusicConditions().add(new MusicCondition(CharmoniumSounds.MUSIC_THARNA, 1200, 3600, mc -> {
+        MusicImprovementsClient.getMusicConditions().add(new MusicCondition(CharmoniumSounds.MUSIC_THARNA, 1200, 3600, mc -> {
             if (mc.player == null || mc.player.world == null)
                 return false;
 
@@ -23,14 +31,14 @@ public class AmbientMusicClient {
         }));
 
         // play Mús in cold environments
-        MusicClient.getMusicConditions().add(new MusicCondition(CharmoniumSounds.MUSIC_MUS, 1200, 3600, mc ->
+        MusicImprovementsClient.getMusicConditions().add(new MusicCondition(CharmoniumSounds.MUSIC_MUS, 1200, 3600, mc ->
             mc.player != null
                 && mc.player.world.getBiome(mc.player.getBlockPos()).getCategory() == Biome.Category.ICY
                 && mc.player.world.random.nextFloat() < 0.28F
         ));
 
         // play Undir in nether underground
-        MusicClient.getMusicConditions().add(new MusicCondition(CharmoniumSounds.MUSIC_UNDIR, 1200, 3600, mc -> {
+        MusicImprovementsClient.getMusicConditions().add(new MusicCondition(CharmoniumSounds.MUSIC_UNDIR, 1200, 3600, mc -> {
             if (mc.player == null)
                 return false;
 
@@ -38,5 +46,9 @@ public class AmbientMusicClient {
                 && DimensionHelper.isDimension(mc.player.world, new Identifier("the_nether"))
                 && mc.player.world.random.nextFloat() < 0.33F;
         }));
+
+        // play Steinn and Draugur in ruins
+        MusicImprovementsClient.getMusicConditions().add(new MusicCondition(CharmoniumSounds.MUSIC_RUIN, 2400, 3600, mc
+            -> mc.player != null && mc.player.world.random.nextFloat() < 0.8F && PlayerStateClient.INSTANCE.ruin));
     }
 }