+++ src/main/java/svenhjol/charmonium/client/ambience/DesertAmbientSounds.java
@@ -4,7 +4,7 @@ import net.minecraft.client.sound.SoundManager;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.sound.SoundEvent;
 import net.minecraft.world.biome.Biome;
-import svenhjol.charm.module.PlayerState;
+import svenhjol.charm.client.PlayerStateClient;
 import svenhjol.charmonium.base.CharmoniumSounds;
 import svenhjol.charmonium.iface.IBiomeAmbience;
 
@@ -34,11 +34,15 @@ public class DesertAmbientSounds {
         }
 
         @Override
+        public float getShortSoundVolume() {
+            return super.getShortSoundVolume() - 0.15F;
+        }
+
+        @Override
         public boolean validBiomeConditions(Biome.Category biomeCategory) {
             return (biomeCategory == Biome.Category.DESERT
-                || biomeCategory == Biome.Category.MESA)
                 && isOutside()
-                && PlayerState.client.isDaytime;
+                && PlayerStateClient.INSTANCE.isDaytime);
         }
     }
 
@@ -67,9 +71,8 @@ public class DesertAmbientSounds {
         @Override
         public boolean validBiomeConditions(Biome.Category biomeCategory) {
             return (biomeCategory == Biome.Category.DESERT
-                || biomeCategory == Biome.Category.MESA)
                 && isOutside()
-                && !PlayerState.client.isDaytime;
+                && !PlayerStateClient.INSTANCE.isDaytime);
         }
     }
 }