+++ src/main/java/svenhjol/charmonium/client/ambience/IcyAmbientSounds.java
@@ -4,7 +4,9 @@ import net.minecraft.client.sound.SoundManager;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.sound.SoundEvent;
 import net.minecraft.world.biome.Biome;
-import svenhjol.charm.module.PlayerState;
+import svenhjol.charm.base.handler.ModuleHandler;
+import svenhjol.charm.client.PlayerStateClient;
+import svenhjol.charm.module.SnowStorms;
 import svenhjol.charmonium.base.CharmoniumSounds;
 import svenhjol.charmonium.iface.IBiomeAmbience;
 
@@ -36,7 +38,7 @@ public class IcyAmbientSounds {
         @Override
         public boolean validBiomeConditions(Biome.Category biomeCategory) {
             return biomeCategory == Biome.Category.ICY
-                && PlayerState.client.isDaytime
+                && PlayerStateClient.INSTANCE.isDaytime
                 && isOutside();
         }
     }
@@ -56,7 +58,32 @@ public class IcyAmbientSounds {
         public boolean validBiomeConditions(Biome.Category biomeCategory) {
             return biomeCategory == Biome.Category.ICY
                 && isOutside()
-                && !PlayerState.client.isDaytime;
+                && !PlayerStateClient.INSTANCE.isDaytime;
+        }
+    }
+
+    public static class Thunderstorm extends BaseAmbientSounds implements IBiomeAmbience {
+        public Thunderstorm(PlayerEntity player, SoundManager soundHandler) {
+            super(player, soundHandler);
+        }
+
+        @Nullable
+        @Override
+        public SoundEvent getLongSound() {
+            return CharmoniumSounds.AMBIENCE_ICY_THUNDERSTORM;
+        }
+
+        @Override
+        public float getLongSoundVolume() {
+            return super.getLongSoundVolume() + 0.25F;
+        }
+
+        @Override
+        public boolean validBiomeConditions(Biome.Category biomeCategory) {
+            return biomeCategory == Biome.Category.ICY
+                && isOutside()
+                && getWorld().isThundering()
+                && ModuleHandler.enabled(SnowStorms.class);
         }
     }
 }