+++ src/main/java/svenhjol/charmonium/client/ambience/MountainsAmbientSounds.java
@@ -4,13 +4,13 @@ import net.minecraft.client.sound.SoundManager;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.sound.SoundEvent;
 import net.minecraft.world.biome.Biome;
-import svenhjol.charm.module.PlayerState;
+import svenhjol.charm.client.PlayerStateClient;
 import svenhjol.charmonium.base.CharmoniumSounds;
 import svenhjol.charmonium.iface.IBiomeAmbience;
 
 import javax.annotation.Nullable;
 
-public class ExtremeHillsAmbientSounds {
+public class MountainsAmbientSounds {
     public static class Day extends BaseAmbientSounds implements IBiomeAmbience {
         public Day(PlayerEntity player, SoundManager soundHandler) {
             super(player, soundHandler);
@@ -19,29 +19,29 @@ public class ExtremeHillsAmbientSounds {
         @Nullable
         @Override
         public SoundEvent getLongSound() {
-            return CharmoniumSounds.AMBIENCE_EXTREME_HILLS_DAY_LONG;
+            return CharmoniumSounds.AMBIENCE_MOUNTAINS_DAY_LONG;
         }
 
         @Nullable
         @Override
         public SoundEvent getShortSound() {
-            return CharmoniumSounds.AMBIENCE_EXTREME_HILLS_DAY_SHORT;
+            return CharmoniumSounds.AMBIENCE_MOUNTAINS_DAY_SHORT;
         }
 
         @Override
         public float getShortSoundVolume() {
-            return super.getShortSoundVolume() - 0.1F;
+            return super.getShortSoundVolume() - 0.15F;
         }
 
         @Override
         public int getShortSoundDelay() {
-            return world.random.nextInt(600) + 500;
+            return world.random.nextInt(600) + 600;
         }
 
         @Override
         public boolean validBiomeConditions(Biome.Category biomeCategory) {
             return biomeCategory == Biome.Category.EXTREME_HILLS
-                && PlayerState.client.isDaytime
+                && PlayerStateClient.INSTANCE.isDaytime
                 && isOutside();
         }
     }
@@ -54,14 +54,14 @@ public class ExtremeHillsAmbientSounds {
         @Nullable
         @Override
         public SoundEvent getLongSound() {
-            return CharmoniumSounds.AMBIENCE_EXTREME_HILLS_NIGHT_LONG;
+            return CharmoniumSounds.AMBIENCE_MOUNTAINS_NIGHT_LONG;
         }
 
         @Override
         public boolean validBiomeConditions(Biome.Category biomeCategory) {
             return biomeCategory == Biome.Category.EXTREME_HILLS
                 && isOutside()
-                && !PlayerState.client.isDaytime;
+                && !PlayerStateClient.INSTANCE.isDaytime;
         }
     }
 }