+++ CHANGELOG.md
@@ -1,4 +1,29 @@
 # Changelog
 
+## 2.3.2
+* Bump for Charm 2.3.2
+
+## 2.3.1
+* Added badlands biome sounds
+* Added more forest, plains and night sounds
+* Added more cave sounds
+
+## 2.2.2
+* Depends on Charm 2.2.2
+
+## 2.2.0
+* Depends on Charm 2.2.0
+* Fix compat issue with Immersive Portals #5
+* Added configurable biome category ambience and dimensions that are considered "outdoors"
+
+## 2.1.9
+* Depends on Charm 2.1.9
+
+## 2.1.4
+* Depends on Charm 2.1.4
+
+## 2.1.3
+* Added tracks for ruins (Quark's big dungeons, Strange's ruins and foundations).
+
 ## 2.0.2
 * Mineshaft and village ambient sounds are less frequent